fix main

+++ movement/movement.h
@@ -1,8 +1,8 @@
 #include <gb/gb.h>
 
 // collision
-extern bool collision_check(UINT8 x1, UINT8 y1, UINT8 w1, UINT8 h1, UINT8 x2, UINT8 y2, UINT8 w2, UINT8 h2);
-extern bool player_collision_with_screen(UINT8 x, UINT8 y, UINT8 player_size, UINT8 tile_size);
+extern bool collision_check(int x1, int y1, int w1, int h1, int x2, int y2, int w2, int h2);
+extern bool player_collision_with_screen(int x, int y, int player_size, int tile_size);
 
 // large
 typedef struct {
